+++ src/frontend/Orion/Sidebar.js
@@ -182,6 +182,10 @@ const ADMIN_SETTINGS_LINKS_CONFIG = [
     label: 'End User Terms Management',
     link: '/orion/configuration/admin-settings/end-user-terms',
   },
+  {
+    label: 'Total History',
+    link: '/orion/configuration/admin-settings/total-history',
+  },
 ]
 
 const renderStyledNavLink = ({ label, link }) => (